refactor: destructured modal-invite-to team
