+++ shared/fragments/modals/src/modal-invite-to-team/modal-invite-to-team.component.tsx
@@ -1,19 +1,8 @@
 import React                       from 'react'
 import { FC }                      from 'react'
-import { FormattedMessage }        from 'react-intl'
-import { useState }                from 'react'
 import { useIntl }                 from 'react-intl'
 
-import { Condition }               from '@ui/condition'
-import { Column }                  from '@ui/layout'
-import { Layout }                  from '@ui/layout'
-import { Modal }                   from '@ui/modal'
-import { Text }                    from '@ui/text'
-import { useGetUser }              from '@shared/data'
-import { useSession }              from '@stores/session'
-
-import { ButtonCreateTeam }        from './button-create-team'
-import { ListTeams }               from './list-teams'
+import { ModalInvite }             from '../modal-invite'
 import { ModalInviteToTeamProps }  from './modal-invite-to-team.interfaces'
 import { useCreateTeamJoinInvite } from './data'
 
@@ -22,25 +11,14 @@ export const ModalInviteToTeam: FC<ModalInviteToTeamProps> = ({
   modalOpen,
   onClose,
 }) => {
-  const [selectedTeam, setSelectedTeam] = useState<string>()
-
   const { formatMessage } = useIntl()
 
   const { createTeamJoiInvite, loading } = useCreateTeamJoinInvite()
-  const { userId } = useSession()
 
-  const { user, teams: allTeams } = useGetUser({ id: userId })
-  const teams = allTeams?.filter((team) => team.ownerUserId === user?.id) || []
-
-  const handleCancel = () => {
-    setSelectedTeam(undefined)
-    onClose?.()
-  }
-
-  const handleSubmit = async () => {
+  const handleSubmit = async (teamId?: string) => {
     try {
       await createTeamJoiInvite({
-        variables: { userId: invitedUserId || '', teamId: selectedTeam || '' },
+        variables: { userId: invitedUserId || '', teamId: teamId || '' },
       })
       onClose?.()
     } catch (error) {
@@ -49,34 +27,12 @@ export const ModalInviteToTeam: FC<ModalInviteToTeamProps> = ({
   }
 
   return (
-    <Modal
-      open={modalOpen}
+    <ModalInvite
       title={formatMessage({ id: 'shared_ui.modal.invite_to_team' })}
-      onClose={handleCancel}
-      onCancel={handleCancel}
-      showCancel={teams.length > 0}
-      showConfirm={teams.length > 0}
-      onConfirm={handleSubmit}
-      confirmText={formatMessage({ id: 'shared_ui.modal.invite' })}
-      confirmProps={{ disabled: !selectedTeam || loading }}
-    >
-      <Condition match={teams.length === 0}>
-        <Column height={120} justifyContent='center' alignItems='center'>
-          <Text fontSize='normal' color='text.gray'>
-            <FormattedMessage id='shared_ui.modal.you_do_not_have_teams' />
-          </Text>
-          <Layout flexBasis={20} flexShrink={0} />
-          <ButtonCreateTeam user={user} />
-        </Column>
-      </Condition>
-      <Condition match={teams.length > 0}>
-        <ListTeams
-          teams={teams}
-          selectedTeam={selectedTeam}
-          onChangeSelectedTeam={setSelectedTeam}
-        />
-        <Layout flexBasis={40} flexShrink={0} />
-      </Condition>
-    </Modal>
+      modalOpen={modalOpen}
+      loading={loading}
+      onClose={onClose}
+      onSubmit={handleSubmit}
+    />
   )
 }
