+++ shared/fragments/modals/src/modal-invite/index.ts
@@ -1,0 +1,2 @@
+export * from './modal-invite.component'
+export * from './modal-invite.interfaces'
