+++ shared/fragments/modals/src/modal-invite/modal-invite.interfaces.ts
@@ -1,0 +1,7 @@
+export interface ModalInviteProps {
+  title?: string
+  modalOpen?: boolean
+  loading: boolean
+  onClose?: () => void
+  onSubmit?: (teamId?: string) => void
+}
